Add files via upload
SQLAlchemy in conjunction with Pandas was used to complete this homework. It must be noted that Pandas', "pd.read_sql" was used as the primary tool to query the SQLLite database, and then the resulting dataframe(s) was(were) worked with to extract the necessary information to complete the different sections of the assignment. Since we were not adding new information to the SQLLite database, I intentionally limited my use of the 'Session' function from SQLAlChemy module.  I was able to complete all sections of this assignment relying on the 'connect', 'pd.read_sql' and other Pandas functions previously used in class.